Move XYZ model/view/template to subdir
