+++ lib/assets/javascripts/cartodb/common/dialogs/add_custom_basemap/add_custom_basemap_model.js
@@ -1,7 +1,6 @@
 var cdb = require('cartodb.js');
-var _ = require('underscore');
 var Backbone = require('backbone');
-var XYZModel = require('./xyz_model.js');
+var XYZModel = require('./xyz/xyz_model.js');
 var WMSModel = require('./wms/wms_model.js');
 
 /**
