+++ lib/assets/javascripts/cartodb/common/dialogs/add_custom_basemap/xyz/xyz_model.js
