+++ lib/assets/javascripts/cartodb/common/dialogs/add_custom_basemap/xyz/xyz_view.js
@@ -18,7 +18,7 @@ module.exports = cdb.core.View.extend({
 
   render: function() {
     this.$el.html(
-      cdb.templates.getTemplate('common/dialogs/add_custom_basemap/xyz')({
+      cdb.templates.getTemplate('common/dialogs/add_custom_basemap/xyz/xyz')({
       })
     );
