+++ lib/assets/test/spec/cartodb/common/dialogs/add_custom_basemap/xyz/xyz_view.spec.js
@@ -1,7 +1,7 @@
-var XYZView = require('../../../../../../javascripts/cartodb/common/dialogs/add_custom_basemap/xyz_view.js');
-var XYZViewModel = require('../../../../../../javascripts/cartodb/common/dialogs/add_custom_basemap/xyz_model.js');
+var XYZView = require('../../../../../../../javascripts/cartodb/common/dialogs/add_custom_basemap/xyz/xyz_view.js');
+var XYZViewModel = require('../../../../../../../javascripts/cartodb/common/dialogs/add_custom_basemap/xyz/xyz_model.js');
 
-describe('common/dialog/add_custom_basemap/xyz_view', function() {
+describe('common/dialog/add_custom_basemap/xyz/xyz_view', function() {
   beforeEach(function() {
     this.baseLayers = new cdb.admin.UserLayers();
     this.model = new XYZViewModel({
